Moved schedule service to its module folder
